Run dockerized pip as invoking user
By default docker executes commands in an container as root. This causes the downloaded dependencies to be owned by root. This messes up cleanup scripts.

Using the correct object which holds the dockerizePip option.
this.serverless.service.custom.dockerizePip

Initialize the this.serverless.service.package.exclude and this.serverless.service.package.include correctly if they are not set.