Translate Pipfiles to Requirements in modules
The plugin does not properly handle Pipfiles that sit in submodules
since the pipenv integration does not anticipate submodules.

This commit reworks the pipenv integration such that submodules are
supported. This changes will convert Pipfiles into requirements as
they are parsed by `pip.js` instead of doing it as a pre-processing
step. This is done since the submodule discovery logic sits in pip.js.