Add support to copy extra files from the Docker image to the package
This allows native OS libraries (`*.so` files) to be copied to the
serverless archive to be available at runtime. This introduces the new
option `dockerExtraFiles`, a list of paths on the Docker image to copy.

For example, the `mysqlclient` package requires
`libmysqlclient.so.1020`. One can now add an item to the
`dockerExtraFiles` option in their `serverless.yml`:

    pythonRequirements:
      dockerExtraFiles:
        - /usr/lib64/mysql57/libmysqlclient.so.1020

This file will be available at runtime in the Lambda function.

This removes the need to manage these files in a more manual way and
avoids the need to commit binary files one's repository.