test: remove parser option from prettier
The `parser` option exists to force prettier to conform to a particular
style. The docs recommend that this only be implemented in an `overrides`
section, to prevent parsing one file as another.
See https://prettier.io/docs/en/configuration.html#setting-the-parserdocsenoptionshtmlparser-option

Since the code is linted correctly without the override, and we want to
lint Markdown, remove this configuration.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>