Add test to check whether the plugin mounts the entire ssh dir
This test makes sure the plugin actually mounts the SSH directory into
the container, if the dockerizePip and dockerSsh are given.
It does this by creating a new known_hosts file and placing the RSA
fingerprint of Github inside, then installing a package via Git, which
requires this fingerprint to be present in known_hosts.
If it is not present, the process will fail.

This patch also adds minor debugging output, in case no
requirements.txt is found.